+++ draft-js-inline-toolbar-plugin/CHANGELOG.md
@@ -7,3 +7,4 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - fixed positioning of the toolbar within a limited space;
 - fixed spontaneous change of toolbar sizes;
+- added external prop `toolbarMargin`;